// src/app/modules/newsEvents/newsEvents.interface.ts
export type TNewsEvents = {
  title: string;
  coverphoto: string;
  date: string;
  source: string;
  description1: string;
  description2: string;
  description3: string;
  description4: string;
  description5: string;
  description6: string;
  description7: string;
  description8: string;
  newsEventSourceLink: string;
  newsOrEvent: "News" | "Events";
  photos: string[];
};